refactor(core): where-parser and rerank move to karajan-core/rag (KJC-TSK-0632)
